# backend/app/docs/swagger.yaml
definitions:
...
  app_internal_handler_tag.CreateTagReq:
    properties:
      color:
        maxLength: 12
        minLength: 3
        type: string
      icon:
        maxLength: 255
        minLength: 3
        type: string
      tag_name:
        example: 工作
        maxLength: 12
        minLength: 1
        type: string
      tag_value:
        example: work
        maxLength: 32
        minLength: 1
        type: string
    required:
    - tag_name
    - tag_value
    type: object
  app_internal_handler_tag.GetTagListResp:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      tags:
        items:
          $ref: '#/definitions/backend_app_types_dto.TagDTO'
        type: array
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  app_internal_handler_tag.UpdateTagReq:
    properties:
      color:
        maxLength: 12
        minLength: 3
        type: string
      icon:
        maxLength: 255
        minLength: 3
        type: string
      tag_name:
        maxLength: 12
        minLength: 1
        type: string
      tag_value:
        maxLength: 32
        minLength: 1
        type: string
    type: object
...
  backend_app_types_dto.TagDTO:
    properties:
      color:
        type: string
      icon:
        type: string
      tag_id:
        type: integer
      tag_name:
        type: string
      tag_value:
        type: string
    type: object
...
  backend_utils_handle.Response:
    properties:
      code:
        description: 响应码，0 表示成功
        example: 0
        type: integer
      data:
        description: 响应数据（可选）
      message:
        description: 响应消息（可选）
        example: 操作成功
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 这是一个基于 Go 和 Gin 框架的 API 服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend API
  version: "1.0"
paths:
...
  /api/tag:
    post:
      consumes:
      - application/json
      description: 创建一个新标签
      parameters:
      - description: 创建标签请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app_internal_handler_tag.CreateTagReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/backend_utils_handle.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_app_types_dto.TagDTO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
      security:
      - BearerAuth: []
      summary: 创建标签
      tags:
      - 标签管理
  /api/tag/{tag_id}:
    delete:
      consumes:
      - application/json
      description: 删除指定标签
      parameters:
      - description: 标签ID
        in: path
        name: tag_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
        "404":
          description: 标签不存在
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
      security:
      - BearerAuth: []
      summary: 删除标签
      tags:
      - 标签管理
    get:
      consumes:
      - application/json
      description: 获取指定标签的详细信息
      parameters:
      - description: 标签ID
        in: path
        name: tag_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/backend_utils_handle.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_app_types_dto.TagDTO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
        "404":
          description: 标签不存在
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
      security:
      - BearerAuth: []
      summary: 获取标签
      tags:
      - 标签管理
    put:
      consumes:
      - application/json
      description: 更新指定标签的信息
      parameters:
      - description: 标签ID
        in: path
        name: tag_id
        required: true
        type: integer
      - description: 更新标签请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app_internal_handler_tag.UpdateTagReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/backend_utils_handle.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_app_types_dto.TagDTO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
        "404":
          description: 标签不存在
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
      security:
      - BearerAuth: []
      summary: 更新标签
      tags:
      - 标签管理
  /api/tag/list:
    get:
      consumes:
      - application/json
      description: 获取标签列表，支持分页
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/backend_utils_handle.Response'
            - properties:
                data:
                  $ref: '#/definitions/app_internal_handler_tag.GetTagListResp'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/backend_utils_handle.Response'
      security:
      - BearerAuth: []
      summary: 获取标签列表
      tags:
      - 标签管理
...
schemes:
- http
- https
swagger: "2.0"